kilka ważnych zmian
1.  Dodanie możliwości przesuwania po płutnie czyli chwytania myszką i przesuwania wykresu w zależności w którym kierunku jest przesuwany kursor,
2. Dodanie przycisku resetowania skali powiększenia/pomniejszenia, oraz centrowania wykresu
3. Dodanie opcji zapisywania wykresów

+++ FuncDraw/FuncDraw/Calculate.cs
@@ -173,19 +173,19 @@ namespace FuncDraw
                         
                             if (tokens[Counter] == "*")
                             {
-                                result = Convert.ToInt64(tokens[Counter - 1]) * Convert.ToInt32(tokens[Counter + 1]);
+                                result = Convert.ToDouble(tokens[Counter - 1]) * Convert.ToDouble(tokens[Counter + 1]);
                             }
                             else if (tokens[Counter] == "/")
                             {
-                                result = Convert.ToInt64(tokens[Counter - 1]) / Convert.ToInt32(tokens[Counter + 1]);
+                                result = Convert.ToDouble(tokens[Counter - 1]) / Convert.ToDouble(tokens[Counter + 1]);
                             }
                             else if (tokens[Counter] == "-")
                             {
-                                result = Convert.ToInt64(tokens[Counter - 1]) - Convert.ToInt32(tokens[Counter + 1]);
+                                result = Convert.ToDouble(tokens[Counter - 1]) - Convert.ToDouble(tokens[Counter + 1]);
                             }
                             else if (tokens[Counter] == "+")
                             {
-                                result = Convert.ToInt64(tokens[Counter - 1]) + Convert.ToInt32(tokens[Counter + 1]);
+                                result = Convert.ToDouble(tokens[Counter - 1]) + Convert.ToDouble(tokens[Counter + 1]);
                             }
                         else if (tokens[Counter] == "^")
                         {
